+++ base_aux/aux_dict/m3_dict_dots1_simple.py
@@ -1,80 +1,11 @@
-from typing import *
+from base_aux.aux_dict.m2_ignorecase import DictIgnorecase
 from base_aux.aux_types.m2_info import ObjectInfo
 
 from base_aux.aux_attr.m2_annot2_nest_required import *
-from base_aux.aux_iter.m1_iter_aux import IterAux
 
 
 # =====================================================================================================================
-class DictCaseinsense(dict):
-    """
-    just a Caseinsense dict
-    """
-    # __getattr__ = dict.get
-    # __setattr__ = dict.__setitem__
-    # __delattr__ = dict.__delitem__
-    # __iter__ = dict.__iter__
-    # __copy__ = dict.copy
-
-    # __repr__ = dict.__repr__  # и так работает!
-    # __str__ = dict.__str__    # и так работает!
-    # __len__ = dict.__len__    # и так работает!
-
-    # -----------------------------------------------------------------------------------------------------------------
-    # GENERIC CLASSES LIKE DICT MUST APPLIED LAST IN MRO!!!
-    # def __init__(self, *args, **kwargs):
-    #     super().__init__(*args, **kwargs)
-
-    # -----------------------------------------------------------------------------------------------------------------
-    def pop(self, item: Any) -> None:
-        item_original = IterAux(self).item__get_original(item)
-        if item_original is None:
-            item_original = item
-
-        super().pop(item_original)
-
-    def get(self, item: Any) -> Any:    # | NoReturn:
-        """
-        always get value or None!
-        if you need check real contain key - check contain)))) [assert key in self] or [getitem_original]
-        """
-        key_original = IterAux(self).item__get_original(item)
-        if key_original is None:
-            return None
-            # key_original = item
-            # msg = f"{item=}"
-            # raise KeyError(msg)
-
-        return super().get(key_original)
-
-    # def set(self, item: Any, value: Any) -> None:
-
-    def update(self, m, /, **kwargs) -> None:
-        for item, value in m.items():
-            key_original = IterAux(self).item__get_original(item)
-            if key_original is None:
-                key_original = item
-
-            super().update({key_original: value})
-
-    def __contains__(self, item: Any) -> bool:
-        if IterAux(self).item__get_original(item) is not None:
-            return True
-
-    # -----------------------------------------------------------------------------------------------------------------
-    # ITEM is universal!
-    def __getitem__(self, item: Any) -> Any | NoReturn:
-        return self.get(item)
-
-    def __setitem__(self, item: Any, value: Any) -> None | NoReturn:
-        self.update({item: value})
-
-    def __delitem__(self, item: Any) -> None:
-        self.pop(item)
-
-
-# =====================================================================================================================
-class DictDots(DictCaseinsense):
+class DictDots(DictIgnorecase):
     """
     dot.notation access to dictionary keys.
     RULES
@@ -228,12 +159,14 @@ class DictDotsAnnotRequired(DictDots, NestInit_AnnotsRequired):
 
     so we cant just call init with super()
     """
+
     def __init__(self, *args, **kwargs) -> None | NoReturn:
         super().__init__(*args, **kwargs)
         # super(NestInit_AnnotsRequired, self).__init__()
         # annot_types = self.annot__get_nested__dict_types()
         # print(f"{annot_types=}")
         self.check_all_defined_or_raise()
+        # FIXME: dont use base from NestInit_AnnotsRequired!!! will not work! or fix!
 
     def check_all_defined_or_raise(self) -> None | NoReturn:
         not_def_list = []