+++ base_aux/aux_dict/m3_dict_dots2_full.py
@@ -1,5 +1,5 @@
 from typing import *
-from base_aux.aux_dict.m2_dict_dots1_simple import *
+from base_aux.aux_dict.m3_dict_dots1_simple import *
 
 
 # =====================================================================================================================