+++ tests/aux_dicts/test__m2_dots_1_simple.py
@@ -1,7 +1,7 @@
 from typing import *
 import pytest
 
-from base_aux.aux_dict.m2_dict_dots1_simple import *
+from base_aux.aux_dict.m3_dict_dots1_simple import *
 
 
 # =====================================================================================================================